==> 01a_Data_Collection_posts.ipynb <==
# Import Libraries

import pandas as pd
import requests
import numpy as np
import time

# Subreddit: Science (Posts)

# This function will collect 100 posts for the subreddit specified 
# and loop over the function 50 times (5000 posts total)

utc = 1610798400
#Jan 16 2021 @ 12:00:00 PM GMT

posts_master = pd.DataFrame(columns = ['subreddit', 'title', 'selftext', 'author', 'created_utc', 'upvote_ratio', 'full_link',
                                       'num_comments', 'num_crossposts', 'total_awards_received', 'score', 'over_18'])

for pull in range(50):
    try:
        query = f'https://api.pushshift.io/reddit/search/submission/?subreddit=science&size=100&before={utc}'
        res = requests.get(query)
        data = res.json()['data']
        pull_dict = {
            'title': [],
            'subreddit': [],
            'created_utc': [],
            'selftext': [],
            'author': [],
            'upvote_ratio': [],
            'full_link': [],            
            'num_comments': [],
            'num_crossposts': [],
            'total_awards_received': [],
            'score': [],
#             'over_18': []
        }
        for i in data:
            pull_dict['title'].append(i['title'])
            pull_dict['subreddit'].append(i['subreddit'])
            pull_dict['created_utc'].append(i['created_utc'])
            pull_dict['selftext'].append(i['selftext'])
            pull_dict['author'].append(i['author'])
            pull_dict['upvote_ratio'].append(i['upvote_ratio'])
            pull_dict['full_link'].append(i['full_link'])
            pull_dict['num_comments'].append(i['num_comments'])
            pull_dict['num_crossposts'].append(i['num_crossposts'])
            pull_dict['total_awards_received'].append(i['total_awards_received'])
            pull_dict['score'].append(i['score'])
#             pull_dict['over_18'].append(i['over_18'])
        temp_posts = pd.DataFrame(pull_dict)
        posts_master = pd.concat([posts_master, temp_posts])
        utc = posts_master['created_utc'].astype('int64').min()
        time.sleep(75)
        print(f'pull number {pull + 1} Complete utc is {utc}') 
    except:
        print('error')
        pass

posts_master.shape

### Export to csv

# science_pull_2 = posts_master
# science_pull_2.to_csv('./data/science_pull_2.csv')

# Subreddit: Technology (Posts)

# This function will collect 100 posts for the subreddit specified 
# and loop over the function 50 times (5000 posts total)

utc = 1610798400
#Jan 16 2021 @ 12:00:00 PM GMT

posts_master = pd.DataFrame(columns = ['subreddit', 'title', 'selftext', 'author', 'created_utc', 'upvote_ratio', 'full_link',
                                       'num_comments', 'num_crossposts', 'total_awards_received', 'score', 'over_18'])

for pull in range(50):
    try:
        query = f'https://api.pushshift.io/reddit/search/submission/?subreddit=technology&size=100&before={utc}'
        res = requests.get(query)
        data = res.json()['data']
        pull_dict = {
            'title': [],
            'subreddit': [],
            'created_utc': [],
#             'selftext': [],
            'author': [],
            'upvote_ratio': [],
            'full_link': [],            
            'num_comments': [],
            'num_crossposts': [],
            'total_awards_received': [],
            'score': [],
#             'over_18': []
        }
        for i in data:
            pull_dict['title'].append(i['title'])
            pull_dict['subreddit'].append(i['subreddit'])
            pull_dict['created_utc'].append(i['created_utc'])
#             pull_dict['selftext'].append(i['selftext'])
            pull_dict['author'].append(i['author'])
            pull_dict['upvote_ratio'].append(i['upvote_ratio'])
            pull_dict['full_link'].append(i['full_link'])
            pull_dict['num_comments'].append(i['num_comments'])
            pull_dict['num_crossposts'].append(i['num_crossposts'])
            pull_dict['total_awards_received'].append(i['total_awards_received'])
            pull_dict['score'].append(i['score'])
#             pull_dict['over_18'].append(i['over_18'])
        temp_posts = pd.DataFrame(pull_dict)
        posts_master = pd.concat([posts_master, temp_posts])
        utc = posts_master['created_utc'].astype('int64').min()
        time.sleep(75)
        print(f'pull number {pull + 1} Complete utc is {utc}') 
    except:
        print('error')
        pass

posts_master.shape

### Export to csv

# technology_pull_2 = posts_master
# technology_pull_2.to_csv('./data/technology_pull_2.csv')
