==> 01b_Data_Collection_comments.ipynb <==
# Import Libraries

import pandas as pd
import requests
import numpy as np
import time

# Subreddit: Science (Comments)

# This function will collect 100 comments for the subreddit specified 
# and loop over the function 50 times (5000 posts total)

utc = 1610798400
#Jan 16 2021 @ 12:00:00 PM GMT

posts_master_science_comments = pd.DataFrame(columns = ['body', 'subreddit', 'created_utc', 'author', 'permalink', 'id', 'num_crossposts', 'total_awards_received', 'score'])
               
for pull in range(50):
    try:
        query = f'https://api.pushshift.io/reddit/search/comment/?subreddit=science&size=100&before={utc}'
        res = requests.get(query)
        data = res.json()['data']
        pull_dict = {
            'created_utc': [],
            'author': [],
            'body': [],
            'subreddit': [],
#             'permalink': [],            
#             'id': [],
#             'num_crossposts': [],
#             'total_awards_received': [],
#             'score': [],
        }
        for i in data:
            pull_dict['created_utc'].append(i['created_utc'])
            pull_dict['author'].append(i['author'])
            pull_dict['body'].append(i['body'])
            pull_dict['subreddit'].append(i['subreddit'])
#             pull_dict['permalink'].append(i['permalink'])
#             pull_dict['id'].append(i['id'])
#             pull_dict['num_crossposts'].append(i['num_crossposts'])
#             pull_dict['total_awards_received'].append(i['total_awards_received'])
#             pull_dict['score'].append(i['score'])
        temp_posts = pd.DataFrame(pull_dict)
        posts_master_science_comments = pd.concat([posts_master_science_comments, temp_posts])
        utc = posts_master_science_comments['created_utc'].astype('int64').min()
        time.sleep(75)
        print(f'pull number {pull + 1} Complete utc is {utc}') 
    except:
        print('error')
        pass

posts_master_science_comments.shape

### Export to csv

# posts_master_science_comments.to_csv('./data/posts_master_science_comments.csv')

# Subreddit: Technology (Comments)

# This function will collect 100 comments for the subreddit specified 
# and loop over the function 50 times (5000 posts total)

utc = 1610798400
#Jan 16 2021 @ 12:00:00 PM GMT

posts_master_technology_comments = pd.DataFrame(columns = ['body', 'subreddit', 'created_utc', 'author', 'permalink', 'id', 'num_crossposts', 'total_awards_received', 'score'])
               
for pull in range(50):
    try:
        query = f'https://api.pushshift.io/reddit/search/comment/?subreddit=technology&size=100&before={utc}'
        res = requests.get(query)
        data = res.json()['data']
        pull_dict = {
            'created_utc': [],
            'author': [],
            'body': [],
            'subreddit': [],
#             'permalink': [],            
#             'id': [],
#             'num_crossposts': [],
#             'total_awards_received': [],
#             'score': [],
        }
        for i in data:
            pull_dict['created_utc'].append(i['created_utc'])
            pull_dict['author'].append(i['author'])
            pull_dict['body'].append(i['body'])
            pull_dict['subreddit'].append(i['subreddit'])
#             pull_dict['permalink'].append(i['permalink'])
#             pull_dict['id'].append(i['id'])
#             pull_dict['num_crossposts'].append(i['num_crossposts'])
#             pull_dict['total_awards_received'].append(i['total_awards_received'])
#             pull_dict['score'].append(i['score'])
        temp_posts = pd.DataFrame(pull_dict)
        posts_master_technology_comments = pd.concat([posts_master_technology_comments, temp_posts])
        utc = posts_master_technology_comments['created_utc'].astype('int64').min()
        time.sleep(75)
        print(f'pull number {pull + 1} Complete utc is {utc}') 
    except:
        print('error')
        pass

posts_master_technology_comments.shape

### Export to csv

# posts_master_technology_comments.to_csv('./data/posts_master_technology_comments.csv')
