==> 02a_Data_Cleaning_posts.ipynb <==
# Read in libraries and Data

import numpy as np
import pandas as pd

science_df = pd.read_csv('./data/raw_data/reddit_science.csv')
technology_df = pd.read_csv('./data/raw_data/reddit_technology.csv')

# Science Data Frame Cleaning

science_df.head(3)

### Null values

science_df.info()

- Mostly null values in 'selftext'
- All null values in over_18

### Drop columns

science_df.drop(columns = 'Unnamed: 0', inplace=True)
science_df.drop(columns = 'over_18', inplace=True)
science_df.drop(columns = 'selftext', inplace=True)

### Value Counts

science_df['subreddit'].value_counts()

science_df['title'].value_counts()

- Possible duplicate posts

science_df['author'].value_counts()

- Noticed a select few authors have submitted a lot of posts

science_df['upvote_ratio'].value_counts()

- most upvotes ratio are 1, possibly not updated as this is a recent dataset

science_df['num_comments'].value_counts()

- most posts have 0 comments

science_df['num_crossposts'].value_counts()

- most posts have 0 crossposts

science_df['total_awards_received'].value_counts()

- most posts have 0 total awards

science_df['score'].value_counts()

- most posts have a score of 1

# Technology Data Cleaning

technology_df.head(3)

### Null Values

technology_df.info()

- Selftext and over_18 have all null values

### Drop Columns

technology_df.drop(columns = 'Unnamed: 0', inplace=True)
technology_df.drop(columns = 'over_18', inplace=True)
technology_df.drop(columns = 'selftext', inplace=True)

### Value Counts

technology_df['subreddit'].value_counts()

technology_df['title'].value_counts()

- Seems to be a large amount of test posts

technology_df['author'].value_counts()

- A few authors have a lot of submitted posts

technology_df['upvote_ratio'].value_counts()

- most posts have upvote ratio of 1

technology_df['num_comments'].value_counts()

- most posts have 0-2 comments

technology_df['num_crossposts'].value_counts()

- most posts have 0 crossposts

technology_df['total_awards_received'].value_counts()

- most posts have 0 awards received

technology_df['score'].value_counts()

- most posts have score of 1

subreddits_df = pd.concat(objs=[science_df, technology_df], axis=0)
subreddits_df.head()

subreddits_df

### Drop duplicates

subreddits_df = subreddits_df.drop_duplicates(subset=['title'])

subreddits_df.shape

science_df = science_df.drop_duplicates(subset=['title'])

science_df.shape

technology_df = technology_df.drop_duplicates(subset=['title'])

technology_df.shape

# Export Clean & Combined Datasets

science_df.to_csv('./data/reddit_science_clean.csv', index=False)

technology_df.to_csv('./data/reddit_technology_clean.csv', index=False)

subreddits_df.to_csv('./data/posts_combined_clean.csv', index=False)
