==> 02b_Data_Cleaning_comments.ipynb <==
# Read in libraries and Data

import numpy as np
import pandas as pd

sci_comments = pd.read_csv('./data/raw_data/posts_master_science_comments.csv')
tech_comments = pd.read_csv('./data/raw_data/posts_master_technology_comments.csv')

# Science Data Frame Cleaning

sci_comments.head(3)

### Null values

sci_comments.info()

### Drop columns

# drop all null columns
sci_comments.drop(columns = 'Unnamed: 0', inplace=True)
sci_comments.drop(columns = 'permalink', inplace=True)
sci_comments.drop(columns = 'id', inplace=True)
sci_comments.drop(columns = 'num_crossposts', inplace=True)
sci_comments.drop(columns = 'total_awards_received', inplace=True)
sci_comments.drop(columns = 'score', inplace=True)

### Value Counts

sci_comments['subreddit'].value_counts()

sci_comments['body'].value_counts()

- Possible duplicate posts

sci_comments['author'].value_counts()

# Technology Data Cleaning

tech_comments.head(3)

### Null Values

tech_comments.info()

### Drop Columns

#drop all null columns
tech_comments.drop(columns = 'Unnamed: 0', inplace=True)
tech_comments.drop(columns = 'permalink', inplace=True)
tech_comments.drop(columns = 'id', inplace=True)
tech_comments.drop(columns = 'num_crossposts', inplace=True)
tech_comments.drop(columns = 'total_awards_received', inplace=True)
tech_comments.drop(columns = 'score', inplace=True)

### Value Counts

tech_comments['subreddit'].value_counts()

tech_comments['body'].value_counts()

- May have duplicate posts

tech_comments['author'].value_counts()

- many [deleted] posts

comments_df = pd.concat(objs=[sci_comments, tech_comments], axis=0)
comments_df.shape

comments_df

### Drop duplicates

comments_df = comments_df.drop_duplicates(subset=['body'])

comments_df.shape

sci_comments = sci_comments.drop_duplicates(subset=['body'])

sci_comments.shape

sci_comments.head(3)

tech_comments = tech_comments.drop_duplicates(subset=['body'])

tech_comments.shape

tech_comments.head(3)

# Export Clean & Combined Datasets

sci_comments.to_csv('./data/science_comments_clean.csv', index=False)

tech_comments.to_csv('./data/technology_comments_clean.csv', index=False)

comments_df.to_csv('./data/comments_combined_clean.csv', index=False)
