==> 04c_Modeling_Decision_Trees.ipynb <==
# Import Libraries and Clean Data

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import MultinomialNB
from sklearn.pipeline import make_pipeline
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split, GridSearchCV, cross_val_score
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics import plot_confusion_matrix, balanced_accuracy_score, f1_score, recall_score, precision_score

from sklearn.tree import DecisionTreeClassifier, plot_tree

import nltk
from nltk.stem import WordNetLemmatizer
from nltk.stem.porter import PorterStemmer
from nltk.tokenize import RegexpTokenizer

subreddits = pd.read_csv('./data/posts_combined_clean.csv')

subreddits.shape

# Decision Tree Model

### X, y, train test split

X = subreddits['title']
y = subreddits['subreddit']

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=123)

### Baseline Model

y.value_counts(normalize=True)

# the target in the dataset for posts was split approximately 57% for r/science.
# This will be the baseline model - if we guessed r/science each time we would 
# be correct 57% of the time

### Modeling

tree_pipe = make_pipeline(
    CountVectorizer(),
    StandardScaler(with_mean=False),
    DecisionTreeClassifier(),
)

tree_pipe.fit(X_train, y_train)

### Model Scoring

preds_tree = tree_pipe.predict(X_test)

precision_tree = precision_score(y_test, preds_tree, pos_label='science')
recall_tree = recall_score(y_test, preds_tree, pos_label='science')
f1_tree = 2 * (precision_tree * recall_tree) / (precision_tree + recall_tree)

print(f'The accuracy training score is    {tree_pipe.score(X_train, y_train)}')
print(f'The accuracy testing score is     {tree_pipe.score(X_test, y_test)}')
print(f'The bac score is                  {balanced_accuracy_score(y_test, tree_pipe.predict(X_test))}')
print(f'The f1_score score is             {f1_tree}')
print(f'The precision is                  {precision_tree}')
print(f'The recall is                     {recall_tree}')

- Baseline decision tree is overfit
- The model fit the training data almost perfectly, but fit the testing data worse.
- The f1 score and balanced accuracy score did worse than the Logistic and Naive Bayes models.

### Decision Tree Grid Search Parameters

tree_pipe = make_pipeline(
    CountVectorizer(),
    StandardScaler(with_mean=False),
    DecisionTreeClassifier(),
)
tree_pipe.fit(X_train, y_train)
print(f'The accuracy training score is    {tree_pipe.score(X_train, y_train)}')
print(f'The accuracy testing score is     {tree_pipe.score(X_test, y_test)}')

params = {
    'decisiontreeclassifier__max_depth': [2, 5, 7, 10, 30],
    'decisiontreeclassifier__min_samples_split': [5, 10, 20, 50],
    'decisiontreeclassifier__min_samples_leaf': [2, 5, 10, 20],
    'decisiontreeclassifier__ccp_alpha': [0.0001, 0.001, 0.01, 0.1, 0, 1, 10],
}

grid = GridSearchCV(tree_pipe, param_grid = params, n_jobs=-1)

grid.fit(X_train, y_train)

grid.best_params_

grid.score(X_train, y_train)

grid.score(X_test, y_test)

grid.best_params_

### Decision Tree with Stop words

tree_pipe = make_pipeline(
    CountVectorizer(stop_words='english'),
    StandardScaler(with_mean=False),
    DecisionTreeClassifier(ccp_alpha=0, max_depth=30, min_samples_leaf=2, min_samples_split=10),
)
tree_pipe.fit(X_train, y_train)
print(f'The accuracy training score is    {tree_pipe.score(X_train, y_train)}')
print(f'The accuracy testing score is     {tree_pipe.score(X_test, y_test)}')
print(f'The bac score is                  {balanced_accuracy_score(y_test, tree_pipe.predict(X_test))}')

### Feature Importance

tree_pipe.named_steps

coefs = tree_pipe.named_steps['decisiontreeclassifier'].feature_importances_

coef_df = pd.DataFrame({'coefs': coefs}, index = tree_pipe.named_steps['countvectorizer'].get_feature_names())
coef_df.nlargest(10, 'coefs')

- The words which had the highest weight on the model included words relating to scientists/research/studies, as well as some tech company names.
